+++ lib/united_states.rb
@@ -2,8 +2,6 @@
 require 'united_states/version'
 require 'united_states/state/designation'
 
-# rubocop: disable Metrics/ModuleLength
-#
 # Top-level namespace for this gem.
 module UnitedStates
   # Thrown when someone attempts to search for a state with
@@ -32,116 +30,73 @@ module UnitedStates
     find_by_postal_code(name_or_postal_code)
   end
 
-  # rubocop: disable Metrics/AbcSize
-  # rubocop: disable Metrics/MethodLength
   # @return [Array<UnitedStates::State::Designation>]
   #  a collection of all U.S. State Designations.
   def self.all
-    [
-      UnitedStates::State::Designation.new(
-        name: 'Alabama', postal_code: 'AL'),
-      UnitedStates::State::Designation.new(
-        name: 'Alaska', postal_code: 'AK'),
-      UnitedStates::State::Designation.new(
-        name: 'Arizona', postal_code: 'AZ'),
-      UnitedStates::State::Designation.new(
-        name: 'Arkansas', postal_code: 'AR'),
-      UnitedStates::State::Designation.new(
-        name: 'California', postal_code: 'CA'),
-      UnitedStates::State::Designation.new(
-        name: 'Colorado', postal_code: 'CO'),
-      UnitedStates::State::Designation.new(
-        name: 'Connecticut', postal_code: 'CT'),
-      UnitedStates::State::Designation.new(
-        name: 'Delaware', postal_code: 'DE'),
-      UnitedStates::State::Designation.new(
-        name: 'Florida', postal_code: 'FL'),
-      UnitedStates::State::Designation.new(
-        name: 'Georgia', postal_code: 'GA'),
-      UnitedStates::State::Designation.new(
-        name: 'Hawaii', postal_code: 'HI'),
-      UnitedStates::State::Designation.new(
-        name: 'Idaho', postal_code: 'ID'),
-      UnitedStates::State::Designation.new(
-        name: 'Illinois', postal_code: 'IL'),
-      UnitedStates::State::Designation.new(
-        name: 'Indiana', postal_code: 'IN'),
-      UnitedStates::State::Designation.new(
-        name: 'Iowa', postal_code: 'IA'),
-      UnitedStates::State::Designation.new(
-        name: 'Kansas', postal_code: 'KS'),
-      UnitedStates::State::Designation.new(
-        name: 'Kentucky', postal_code: 'KY'),
-      UnitedStates::State::Designation.new(
-        name: 'Louisiana', postal_code: 'LA'),
-      UnitedStates::State::Designation.new(
-        name: 'Maine', postal_code: 'ME'),
-      UnitedStates::State::Designation.new(
-        name: 'Maryland', postal_code: 'MD'),
-      UnitedStates::State::Designation.new(
-        name: 'Massachusetts', postal_code: 'MA'),
-      UnitedStates::State::Designation.new(
-        name: 'Michigan', postal_code: 'MI'),
-      UnitedStates::State::Designation.new(
-        name: 'Minnesota', postal_code: 'MN'),
-      UnitedStates::State::Designation.new(
-        name: 'Mississippi', postal_code: 'MS'),
-      UnitedStates::State::Designation.new(
-        name: 'Missouri', postal_code: 'MO'),
-      UnitedStates::State::Designation.new(
-        name: 'Montana', postal_code: 'MT'),
-      UnitedStates::State::Designation.new(
-        name: 'Nebraska', postal_code: 'NE'),
-      UnitedStates::State::Designation.new(
-        name: 'Nevada', postal_code: 'NV'),
-      UnitedStates::State::Designation.new(
-        name: 'New Hampshire', postal_code: 'NH'),
-      UnitedStates::State::Designation.new(
-        name: 'New Jersey', postal_code: 'NJ'),
-      UnitedStates::State::Designation.new(
-        name: 'New Mexico', postal_code: 'NM'),
-      UnitedStates::State::Designation.new(
-        name: 'New York', postal_code: 'NY'),
-      UnitedStates::State::Designation.new(
-        name: 'North Carolina', postal_code: 'NC'),
-      UnitedStates::State::Designation.new(
-        name: 'North Dakota', postal_code: 'ND'),
-      UnitedStates::State::Designation.new(
-        name: 'Ohio', postal_code: 'OH'),
-      UnitedStates::State::Designation.new(
-        name: 'Oklahoma', postal_code: 'OK'),
-      UnitedStates::State::Designation.new(
-        name: 'Oregon', postal_code: 'OR'),
-      UnitedStates::State::Designation.new(
-        name: 'Pennsylvania', postal_code: 'PA'),
-      UnitedStates::State::Designation.new(
-        name: 'Rhode Island', postal_code: 'RI'),
-      UnitedStates::State::Designation.new(
-        name: 'South Carolina', postal_code: 'SC'),
-      UnitedStates::State::Designation.new(
-        name: 'South Dakota', postal_code: 'SD'),
-      UnitedStates::State::Designation.new(
-        name: 'Tennessee', postal_code: 'TN'),
-      UnitedStates::State::Designation.new(
-        name: 'Texas', postal_code: 'TX'),
-      UnitedStates::State::Designation.new(
-        name: 'Utah', postal_code: 'UT'),
-      UnitedStates::State::Designation.new(
-        name: 'Vermont', postal_code: 'VT'),
-      UnitedStates::State::Designation.new(
-        name: 'Virginia', postal_code: 'VA'),
-      UnitedStates::State::Designation.new(
-        name: 'Washington', postal_code: 'WA'),
-      UnitedStates::State::Designation.new(
-        name: 'West Virginia', postal_code: 'WV'),
-      UnitedStates::State::Designation.new(
-        name: 'Wisconsin', postal_code: 'WI'),
-      UnitedStates::State::Designation.new(
-        name: 'Wyoming', postal_code: 'WY')
-    ]
+    array_from_yaml_file(path: config_path)
   end
-  # rubocop: enable Metrics/AbcSize
-  # rubocop: enable Metrics/MethodLength
+
+  # @example
+  #  UnitedStates.array_from_hashes(
+  #    { name: 'Florida', postal_code: 'FL' },
+  #    { name: 'Iowa', postal_code: 'IA'})
+  # @param hashes [Array<Hash>]
+  #  a collection of attributes as hashes
+  # @return [Array<UnitedStates::State::Designation>]
+  #  a collection of state designations made from the given hash attributes
+  def self.array_from_hashes(*hashes)
+    hashes.map do |hash|
+      UnitedStates::State::Designation.from_hash(hash)
+    end
+  end
+
+  # @example
+  #  require 'united_states'
+  #
+  #  UnitedStates.array_from_yaml(
+  #    { Washington: { postal_code: 'WA' } }.to_yaml)
+  #  # => [<UnitedStates::State::Designation...>]
+  # @param yaml [String]
+  #  a String representation of YAML.
+  # @return [Array<UnitedStates::State::Designation>]
+  #  a collection of state designations made from the given hash attributes
+  # @raise UnitedStates::State::PostalCode::StringTooLongError
+  #  if a designation has a postal code that too long
+  # @raise UnitedStates::State::PostalCode::StringTooShortError
+  #  if a designation has no postal code defined or is too short
+  def self.array_from_yaml(yaml)
+    return [] unless YAML.safe_load(yaml)
+    YAML.safe_load(yaml).map do |key, value|
+      UnitedStates::State::Designation.new(
+        name: key, postal_code: value && value.fetch('postal_code', ''))
+    end
+  end
+
+  # @example
+  #  require 'united_states'
+  #
+  #  UnitedStates.array_from_yaml_file(path: './states.yml')
+  #  # => [<UnitedStates::State::Designation...>]
+  # @param path [String]
+  #  the path to the YAML file
+  # @return [Array<UnitedStates::State::Designation>]
+  #  a collection of state designations made from the YAML file
+  # @raise UnitedStates::State::PostalCode::StringTooLongError
+  #  if a designation has a postal code that too long
+  # @raise UnitedStates::State::PostalCode::StringTooShortError
+  #  if a designation has no postal code defined or is too short
+  def self.array_from_yaml_file(path:)
+    pathname = Pathname.new(path)
+    return array_from_yaml(pathname.read) if pathname.exist?
+    raise "\"#{path}\" does not exist.\n"\
+          'Please supply a path to a YAML file.'
+  end
+
+  # @return [String]
+  #  the path to the Designations yaml file
+  def self.config_path
+    './lib/united_states/designations.yml'
+  end
 
   # @example
   #  UnitedStates.find_by_name('louisiana') # => UnitedSt...Designation